capredv2.backend.domain.tests/DomainEntities/Dropdowns/MilestonesDTOTests.cs
using capredv2.backend.domain.DatabaseEntities.Dropdowns;
using capredv2.backend.domain.DomainEntities.Dropdowns;
using NUnit.Framework;
using System;
using System.Collections.Generic;
using System.Text;

namespace capredv2.backend.domain.tests.DomainEntities.Dropdowns
{
    [TestFixture]
    public class MilestonesDTOTests
    {
        [Test]
        public void MapFromDomainEntity_ValidEntity_ReturnDTOEntity()
        {
            //Arrange
            var milestones = new Milestones()
            {
                Id = Guid.NewGuid(),
                Value = "AMER",
                Position = 0
            };
            var response = MilestonesDTO.MapFromDatabaseEntity(milestones);

            Assert.IsNotNull(response);
            Assert.AreEqual(milestones.Id, response.Id);
            Assert.AreEqual(milestones.Value, response.Value);
            Assert.AreEqual(milestones.Position, response.Position);
        }

        [Test]
        public void MapFromDomainEntity_NullContent_ReturnNull()
        {
            //Act
            var response = MilestonesDTO.MapFromDatabaseEntity(null);

            //Assert
            Assert.IsNull(response);
        }
    }
}
